+++ first_app/hello.py
@@ -7,7 +7,6 @@ from werkzeug.security import check_password_hash, generate_password_hash
 from functools import wraps
 from flask import g, redirect, url_for
 from .models import Post, User
-from .db import get_db, db_session
 from . import db
 
 # app = create_app()
